+++ app/src/src_nonfreenet/org/breezyweather/sources/polleninfo/PollenInfoService.kt
@@ -18,7 +18,6 @@ package org.breezyweather.sources.polleninfo
 
 import android.content.Context
 import breezyweather.domain.location.model.Location
-import breezyweather.domain.source.SourceContinent
 import breezyweather.domain.source.SourceFeature
 import breezyweather.domain.weather.model.Pollen
 import breezyweather.domain.weather.wrappers.PollenWrapper
@@ -27,18 +26,10 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import io.reactivex.rxjava3.core.Observable
 import org.breezyweather.BuildConfig
 import org.breezyweather.R
-import org.breezyweather.common.extensions.currentLocale
-import org.breezyweather.common.extensions.getCountryName
 import org.breezyweather.common.extensions.toCalendarWithTimeZone
 import org.breezyweather.common.extensions.toTimezoneSpecificHour
 import org.breezyweather.common.preference.EditTextPreference
 import org.breezyweather.common.preference.Preference
-import org.breezyweather.common.source.ConfigurableSource
-import org.breezyweather.common.source.HttpSource
-import org.breezyweather.common.source.PollenIndexSource
-import org.breezyweather.common.source.WeatherSource
-import org.breezyweather.common.source.WeatherSource.Companion.PRIORITY_HIGH
-import org.breezyweather.common.source.WeatherSource.Companion.PRIORITY_NONE
 import org.breezyweather.domain.settings.SourceConfigStore
 import org.breezyweather.sources.polleninfo.json.PollenContamination
 import org.breezyweather.sources.polleninfo.json.PollenInfoResult
@@ -52,52 +43,15 @@ import javax.inject.Named
 class PollenInfoService @Inject constructor(
     @ApplicationContext context: Context,
     @Named("JsonClient") client: Retrofit.Builder,
-) : HttpSource(), WeatherSource, PollenIndexSource, ConfigurableSource {
+) : PollenInfoServiceStub(context) {
 
-    override val id = "polleninfo"
-    private val countryName = context.currentLocale.getCountryName("AT")
-    override val name = "Pollen Information Service".let {
-        if (it.contains(countryName)) {
-            it
-        } else {
-            "$it ($countryName)"
-        }
-    }
-    override val continent = SourceContinent.EUROPE
     override val privacyPolicyUrl = "https://www.polleninformation.at/en/our-terms-of-use"
 
     private val mPollenApi by lazy {
         client
-            .baseUrl(POLLENFINO_BASE_URL)
+            .baseUrl(POLLENINFO_BASE_URL)
             .build()
             .create(PollenInfoApi::class.java)
-    }
-
-    private val weatherAttribution = "Österreichischer Polleninformationsdienst (Creative Commons Attribution 4.0)"
-    override val supportedFeatures = mapOf(
-        SourceFeature.POLLEN to weatherAttribution
-    )
-    override val attributionLinks = mapOf(
-        "Österreichischer Polleninformationsdienst" to "www.polleninformation.at"
-    )
-
-    override fun isFeatureSupportedForLocation(
-        location: Location,
-        feature: SourceFeature,
-    ): Boolean {
-        return SUPPORTED_COUNTRY_CODES.any { it.equals(location.countryCode, ignoreCase = true) }
-    }
-
-    override fun getFeaturePriorityForLocation(
-        location: Location,
-        feature: SourceFeature,
-    ): Int {
-        return when {
-            feature == SourceFeature.POLLEN &&
-                location.countryCode.equals("AT")
-            -> PRIORITY_HIGH
-            else -> PRIORITY_NONE
-        }
     }
 
     override fun requestWeather(
@@ -143,75 +97,38 @@ class PollenInfoService @Inject constructor(
         )
     }
 
+    /**
+     * @param contaminationList
+     * @param dayNumber: 1 for today (contamination_1), 2 for tomorrow (contamination_2), etc.
+     */
     private fun getPollenForDay(
         contaminationList: List<PollenContamination>?,
-        dayNumber: Int, // 1 for today (contamination_1), 2 for tomorrow (contamination_2), etc.
+        dayNumber: Int,
     ): Pollen? {
         if (contaminationList.isNullOrEmpty()) {
             return null
         }
 
-        fun PollenContamination.getContaminationForDay(day: Int): Int? =
-            when (day) {
-                1 -> contamination1
-                2 -> contamination2
-                3 -> contamination3
-                4 -> contamination4
-                else -> null
-            }
-
-        var alder: Int? = null
-        var birch: Int? = null
-        var cypress: Int? = null
-        var grass: Int? = null
-        var hazel: Int? = null
-        var mugwort: Int? = null
-        var olive: Int? = null
-        var plane: Int? = null
-        var ragweed: Int? = null
-        var urticaceae: Int? = null
-        var mold: Int? = null
-
-        contaminationList.forEach { contamination ->
-            val value = contamination.getContaminationForDay(dayNumber)
-            if (value != null) {
-                when (contamination.pollId) {
-                    1 -> alder = value // alder (Alnus)
-                    2 -> birch = value // birch (Betula)
-                    3 -> hazel = value // hazel (Corylus)
-                    5 -> grass = value // grasses (Poaceae)
-                    6 -> ragweed = value // ragweed (Ambrosia)
-                    7 -> mugwort = value // mugwort (Artemisia)
-                    15 -> urticaceae = value // nettle family (Urticaceae)
-                    16 -> plane = value // plane tree (Platanus)
-                    17 -> cypress = value // cypress family (Cupressaceae)
-                    18 -> olive = value // olive (Olea)
-                    23 -> mold = value // fungal spores (Alternaria)
-                    //  291 -> rye = value // rye (Secale) - does not exist in BW, could be part of grasses
-                }
-            }
-        }
-
         return Pollen(
-            alder = alder?.pollenIndex,
-            birch = birch?.pollenIndex,
-            cypress = cypress?.pollenIndex,
-            grass = grass?.pollenIndex,
-            hazel = hazel?.pollenIndex,
-            mugwort = mugwort?.pollenIndex,
-            olive = olive?.pollenIndex,
-            plane = plane?.pollenIndex,
-            ragweed = ragweed?.pollenIndex,
-            urticaceae = urticaceae?.pollenIndex,
-            mold = mold?.pollenIndex
+            alder = contaminationList.firstOrNull { it.pollId == 1 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            birch = contaminationList.firstOrNull { it.pollId == 2 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            cypress = contaminationList.firstOrNull { it.pollId == 17 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            grass = contaminationList.firstOrNull { it.pollId == 5 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            hazel = contaminationList.firstOrNull { it.pollId == 3 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            mugwort = contaminationList.firstOrNull { it.pollId == 7 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            olive = contaminationList.firstOrNull { it.pollId == 18 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            plane = contaminationList.firstOrNull { it.pollId == 16 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            ragweed = contaminationList.firstOrNull { it.pollId == 6 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            urticaceae = contaminationList.firstOrNull { it.pollId == 15 }
+                ?.getContaminationForDay(dayNumber)?.pollenIndex,
+            // TODO: Replace with Alternaria:
+            mold = contaminationList.firstOrNull { it.pollId == 23 }?.getContaminationForDay(dayNumber)?.pollenIndex,
+            // TODO: Add rye (Secale)
+            // rye = contaminationList.firstOrNull { it.pollId == 291 }?.getContaminationForDay(dayNumber)?.pollenIndex,
         )
     }
 
-    override val pollenLabels = R.array.pollen_levels_polleninfo
-    override val pollenColors = R.array.pollen_level_colors_polleninfo
-
     // CONFIG
-
     private val config = SourceConfigStore(context, id)
     private var apikey: String
         set(value) {
@@ -224,8 +141,6 @@ class PollenInfoService @Inject constructor(
 
     override val isConfigured
         get() = getApiKeyOrDefault().isNotEmpty()
-
-    override val isRestricted = true
 
     override fun getPreferences(context: Context): List<Preference> {
         return listOf(
@@ -245,21 +160,6 @@ class PollenInfoService @Inject constructor(
     }
 
     companion object {
-        private const val POLLENFINO_BASE_URL = "https://www.polleninformation.at/"
-        private val SUPPORTED_COUNTRY_CODES: List<String> = listOf(
-            "AT", // Austria
-            "CH", // Switzerland
-            "DE", // Germany
-            "ES", // Spain
-            "FR", // France
-            "GB", // United Kingdom
-            "IT", // Italy
-            "LT", // Lithuania
-            "LV", // Latvia
-            "PL", // Poland
-            "SE", // Sweden
-            "TR", // Turkey
-            "UA" // Ukraine
-        )
+        private const val POLLENINFO_BASE_URL = "https://www.polleninformation.at/"
     }
 }